riscv: don't read riscv,isa in the riscv cpu's get_desc() & fix extension detection
riscv,isa is not a property which can be assumed always to be present,
the new "riscv,isa-extensions" and "riscv,isa-base" properties may be
present instead. //TODO fix this fn: cpu_get_desc() returns only a
string, so rather than add support for the list of strings method, just
return the compatible as the cpu description instead.

This patch breaks extension support detection since the returned string
has not of the same meaning, but it was already fundamentally broken:
The get_desc() code returns an error if the isa string is longer than
the provided buffer, and the provided buffer is only 32 elements long.
Many ISA strings are a lot longer.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>