board: mpfs_icicle: implement board_fdt_blob_setup()
The firmware on the Icicle is capable of providing a devicetree in a1 to
U-Boot, but until now the devicetree has been packaged in a "payload" [1]
alongside U-Boot (or other bootloaders/RTOSes) and appended to the image.
The address of this appended devicetree is placed in a1 by the firmware.
This meant that the mechanism used by OF_SEPARATE to locate the
devicetree at the end of the image would pick up the one provided by the
firmware when u-boot-nodtb.bin was in the payload and U-Boot's devicetree
when u-boot.bin was.

The firmware is now going to be capable of providing a minimal devicetree
(quite cut down due to severe space constraints), but this devicetree is
linked into the firmware that runs out of the L2 rather than at the end
of the U-Boot image. Implement board_fdt_blob_setup() so that this
devicetree can be optionally used, and the devicetree provided in the
"payload" can be used without relying on "happening" to implement the
same strategy as OF_SEPARATE expects in combination with
u-boot-nodtb.bin. Unlike other RISC-V boards, the firmware provided
devicetree is only used when OF_BOARD is set, so that the almost
certainly more complete devicetree in U-Boot will be used unless
explicitly requested otherwise.

Link: https://github.com/polarfire-soc/hart-software-services/blob/master/tools/hss-payload-generator/README.md [1]
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>